correction des erreurs

+++ SAE2_02/src/Arc.java
@@ -2,7 +2,7 @@ public class Arc {
     private String dest;
     private double cout;
 
-    public Arc(Noeud n, int cout) {
+    public Arc(Noeud n, double cout) {
         this.dest = n.getNom();
         if (cout>=0){
             this.cout = cout;
